Allow RELEASE_NAME to be overridden in environment (#1087)
RELEASE_NAME was locked to a static value which prevented multiple
instances of Firezone from running in the same network namespace
(i.e. using net=host in a container). This change uses the value of
the RELEASE_NAME environment variable and defaults to the current
static value if not set.