Allow btree_gist extension to preexist
This commit has the exact same rationale as cc280dae, just for a
different extension that I unfortunately missed the first time around.

This time I searched the repo for occurrences of `CREATE EXTENSION`
instead of the name of the extension causing the error, so hopefully I
did not miss anything else.