feat: add phoenix channel for communication with the gateway
This commit adds the communication channel with the gateway and all the
logic surrounding its registrations and connection.

We remove the pg notifier logic so we just hook into the precise moment
the event happens in elixir and send that event to the gateway.

Each gateway connects both to an "all" channel and a channel given by
its id. A gateway can only connect to a channel for its given ID based
on the provided JWT.

The default gateway is created when it registers for the first time if the
gateway doesn't yet exists it creates it. This has the disadvantage of
making the token for registration non-expirable however it works for now
until we have a UI to create gateways.