Expose some http client ssl opts via HTTP_CLIENT_SSL_OPTS
Expose the most commonly-used SSL client options to our
OIDC and ConnectivityChecks HTTP clients.

SSL misconfiguration can be a security concern, so we
intentionally puke when an unexpected key is passed. This
should result in a new GitHub issue being opened and dialog
created to learn more about the use-case.

Fixes #996