Make default IP range CGNAT; randomize IP numbering (#1200)
~~Allocating IP addresses for device creation is not a trivial
endeavor.~~

~~Before, we were using a window function and full-table lock in order
to select a candidate device IP to use when creating a new device. This
behavior was designed to mimic DHCP.~~

~~Consider the following example.~~

~~Say you currently have a list of devices in the database with the
following IPs:~~

~~10.0.0.1
10.0.0.3
10.0.0.100
10.0.0.101~~

~~Assuming WIREGUARD_IPV4_NETWORK is set to 10.0.0.0/24, the IP address
for the next device generated _should_ be 10.0.0.2. Before, this IP
would have been found with a large SELECT query that tried to find the
first non-sequential IP address in the devices table. This required
wrapping the SELECT and following INSERT in a single transaction to
ensure the IP address "reserved" would be insertable.~~

~~As you may imagine, this resulted in deadlocks, and prevented us from
running much of the device tests with `async: true` enabled.~~

~~@conectado came up with a simple but elegant solution to this problem
using a Mutex to guard an IP allocation cache. It works as follows:~~

~~1. Maintain an allocation cache: a list of all of the available IP
    addresses represented as a list of {gap_start, gap_end} tuples.
    Using the example data above, this would be:
    ```
    [
      {10.0.0.2, 10.0.0.2},
      {10.0.0.4, 10.0.0.99},
      {10.0.0.102, 10.0.0.254}
    ]
    ```
2. When creating a new device, lock the mutex, then select the first
    available IP address from the allocation cache. Increment the IP
    address and save this new state as the new allocation cache.
3. If allocation fails (i.e. the cache is empty), rebuild the cache from
    the DB. This has the nice effect that a stale cache just skips a few
    IPs and causes a premature cache rebuild, but doesn't otherwise halt
    creation like a `SEQUENCE` field would.
4. If the rebuilt cache is **also** empty, the IP address pool actually
    *has* been exhausted, so return `nil` for the IP. The Device schema
interprets this as an error and adds an appropriate error to the
changeset.~~

**EDIT**: After further research into Postgres gapless sequences and the
information theory realities the problem is bound by, I'm not so sure
creating a global mutex to maintain strict consistency for sequential IP
allocation is the right approach.

Going to explore some simpler methods of IP allocation, likely involving
random selection inside the CGNAT range.

Co-authored-by: Andrew Dryga <andrew@dryga.com>