API Token CLI (#1271)
Adds a mechanism for generating API tokens from the CLI. Requires the
default admin user to be present. From there the token can be used to
create additional admins. In the future, we could allow specifying a
user's email to generate the token for.

Generate like so:

```
docker compose run --rm firezone bin/create-api-token
```