Disable `on_demand` cert provisioning for Caddy (#1720)
As described in the Caddy documentation, `on_demand` [must be
restricted](https://caddyserver.com/docs/automatic-https#using-on-demand-tls)
to prevent abuse. In our case, **it was not**, leading to DoS vector in
which a malicious client could DoS the `caddy` service by repeatedly
initiating TLS requests with invalid domains at a rate high enough for
the ACME service to block the Firezone server.

Verified the new caddy configuration works as expected.