Xcode: Use Build Phase script to copy connlib headers
To copy the Connlib headers and Swift files, we now use a Build Phase in the
network extension target (instead of a separate target by itself). This enables
us to speficy an output files list. Without this, Xcode builds fail sometimes,
because it expects the source files to be present at the start of the build.