feat(relay): add some basic prometheus metrics (#1742)
The metrics are available at `http://{listen_addr}:8080/metrics`.
Currently, we collect the following:

- Number of active allocations: We can have an alert once the number of
allocations passes through a certain threshold.
- Outcome (success / error) and message kind (allocation / channel_bind
/ ...) of all responses: Summing all of these up would give you the
total number of requests handled. We might want to have a grafana alert
for an increased number of error responses.
- Total number of bytes relayed: Dividing this by time gives us an
average "internal" bandwidth.

This is just a start, we can explore what else is useful as we have it
operate.

Depends-On: https://github.com/firezone/firezone/pull/1743