feat(relay): support IPv6 allocations (#1814)
This patch series adds support for IPv6 allocations. If not specified
otherwise in the ALLOCATE request, clients will get an IP4 allocation.
They can also request an IPv6 address or an additional IPv6 address in
addition to their IPv4 address.

Either of those is only possible if the relay actually has a listening
socket for the requested address family. The CLI is designed such that
the user can either specify IP4, IP6 or both of them.

The `Server` component handles all of this logic and responds with
either a successful allocation response or an Address Family Not
Supported error (see
https://www.rfc-editor.org/rfc/rfc8656#name-stun-error-response-codes).

Multiple refactorings were necessary to achieve this design, they are
all extracted into separate PRs:

Depends-On: #1831.
Depends-On: #1832.
Depends-On: #1833.

---------

Co-authored-by: Jamil <jamilbk@users.noreply.github.com>