fix(relay): wait for channel binding in smoke test script (#1853)
`webrtc-rs` has a race condition where `send_to` does not actually await
the channel binding, thus attempting to send something through the
channel from the other end my fail because we receive the bytes from the
relay before the library registers that there is an active channel.

This should hopefully fix the flakiness of the smoke test script.