Use explicit return codes in JNI method signatures
This fixes a failure that was occurring silently because the JNI method
signatures did not have their return codes included. The `call_method`
function calls fail, and no error is returned when this happens. For
reference on how to specify JNI method signatures:

https://docs.oracle.com/javase/8/docs/technotes/guides/jni/spec/types.html

The FFI callbacks are working now, but we're seeing another crash
related to concurrency between Rust and the JVM which will be dealt
with in another PR.