Apple: Add support for magic links (#1909)
Addresses #1899.

Tested only with PortalMock. Works with Safari. Not tested yet with
non-Safari browsers as default in macOS (if required, will address that
separately).

`client_csrf_token` is always passed. It's verified only if the sign in
happens with the external open-app-with-URL scenario. It's not checked
if the user logs in inside of the `ASWebAuthenticationSession` webpage
itself.