Introduce token uninitialized state
Previously, when authResponse is nil, we consider is as logged out.

This is a problem because keychain access takes time, so objects are
initialized and we don't yet have the authResponse from the keychain.
This is treated as logged out state, and the tunnel is stopped.

So we create a LoginStatus enum which has an uninitialized state.