Fix JNI callbacks thread safety. tl;dr: use GlobalRef
A JObject with static lifetime was incorrectly used to maintain a
reference to the callback class instance in the Java VM. This pointer
was being GC'd, causing an "invalid jobject frame reference" exception
to occur.

The fix, as clearly laid out in the [JNI crate docs
here](https://docs.rs/jni/latest/jni/objects/struct.GlobalRef.html),
is to obtain a global reference to the callback JObject which is
guaranteed by the VM not to be GC'd until explicitly freed.