fix(apple): Don't close fd on `disconnect` (#1965)
When path.status changes to `.unsatisfied`, we call `disconnect()`. This
was closing the tunnel `fd`, causing the PacketTunnelProvider interface
to go down. Then, when `connect()` was being called later, the `fd` was
still closed.

This PR simply removes the `close(fd)` call and assumes the
PacketTunnelProvider will do that appropriately.


After some more discussion with @conectado we determined that boringtun
and webrtc-rs probably handle the edge cases needed by `bumpSockets` and
`disableSomeRoamingForBrokenMobileSemantics`, so those are removed as
well.

Refs firezone/product#656