Connlib/fix stability issues (#1974)
When we lost networks(or change them), the phoenix channel didn't detect
that the connection was lost, since the underlying websocket doesn't
return an error if it's not closed gracefully.

So we expect the heartbeat at some point to consider the connection
down.

Furthermore, while the connection is down sending the connection intents
to the portal fails silently, so now we re-try the message until we get
a response and built some race-condition protections in case we get
multiple or stale responses.