Split magic link token into two parts: emailed and stored in session
This allows to prevent attacks on magic links by emailing the user a fake magic link (to a wrong account) at the moment when he expects a real one and which will result into connecting to a faux gateway that can log traffic.

With this fix there is no way for the attacker to persist the secret part to the browser cookie so the link will be just invalid.