feat(relay): better integration with google cloud trace (#1994)
The relay already uses spans for its internal logging but up until now,
those were only useful when using the regular stdout logger. For
`JSON_LOG=true`, we did emit the spans but not in a format that worked
with Google Cloud Trace.

We are changing the configuration of the binary to remove `JSON_LOG` and
instead configure it with the Google Cloud "project id". If that one is
set, we configure the logs such that they properly integrate with Google
Cloud Trace by emitting unique span IDs for all messages. This should
give us much better observability throughout a request that is being
processed by the relay.

The `JSON_LOG` configuration option has been bugging me for a while
because it was implicitly (a little bit) geared towards Google
Stackdriver. We now make this explicit by requiring the project ID. That
is a much cleaner integration IMO.