fix(relay): don't panic if waking time is in the past (#2052)
To be resource efficient, the relay's event loop suspends if there is no
activity on its sockets. Certain operations however need to happen at a
specified time, for example, allocations need to be freed if not renewed
after a specified TTL. To achieve this, the relay implementation has a
mechanism where it can request the event-loop to wake up at a specified
deadline.

Previously, we assumed that this deadline was always in the future. For
reasons not yet known, we encountered a situation where this deadline
was in the past.

We fix this by:

1) Scheduling an immediate wake-up instead of panicking
2) Logging a warning to help with debugging if this ever occurs again