feat(relay): respond to bad STUN message where possible (#2073)
Previously, we would only log an error if we failed to decode the
attribute part of a STUN message. Now, if we can decode the STUN header
but fail at one of the attributes, we will properly respond to the
client.

This could for example happen if a client requests a channel number that
is out of range.