Install and cache Rust toolchain (#2353)
On GH-hosted runners, this isn't a problem. But on self-hosted runners
with persistent disks, we hit this issue:
https://github.com/Swatinem/rust-cache#known-issues

This PR should fix that by ensuring the Rust toolchain is installed. It
also installs the toolchain after the cache has been restored so we can
take advantage of any cached artifacts.